# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registro de usuário!",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "JWT token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user-profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Perfil dos usuários!",
                "operationId": "78c9456d1e8297ec68dee375323edd5d",
                "responses": {
                    "200": {
                        "description": "Sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/search/local/{cep}": {
            "get": {
                "tags": [
                    "Cep-API"
                ],
                "summary": "Consulta de múltiplos endereços.",
                "description": "Retorna informações de endereço para os CEP(s) fornecidos. Aceita vários CEPs separados por vírgulas.",
                "operationId": "de7848c20f42389b4e6103525693bedf",
                "parameters": [
                    {
                        "name": "cep",
                        "in": "path",
                        "description": "Um único ou mútiplos ceps.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "01001000,17560-246"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "cep": {
                                                "type": "string",
                                                "example": "01001-000"
                                            },
                                            "logradouro": {
                                                "type": "string",
                                                "example": "Praça da Sé"
                                            },
                                            "bairro": {
                                                "type": "string",
                                                "example": "Sé"
                                            },
                                            "localidade": {
                                                "type": "string",
                                                "example": "São Paulo"
                                            },
                                            "uf": {
                                                "type": "string",
                                                "example": "SP"
                                            },
                                            "unidade": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "ibge": {
                                                "type": "string",
                                                "example": "3550308"
                                            },
                                            "gia": {
                                                "type": "string",
                                                "example": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Formato de CEP inválido!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Mensagem": {
                                            "type": "string",
                                            "example": "Formato de CEP inválido! Os CEPs devem conter 8 dígitos contínuos ou cinco dígitos contínuos separados por - seguidos de três outros dígitos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Cep-API",
            "description": "Cep-API"
        }
    ]
}